refactor: 필터 설정 클래스 위치 수정 (#51)
