+++ src/main/kotlin/kr/wooco/woocobe/common/ui/config/WebFilterConfig.kt
@@ -1,12 +1,14 @@
-package kr.wooco.woocobe.common.ui.logging
+package kr.wooco.woocobe.common.ui.config
 
+import kr.wooco.woocobe.common.ui.filter.MdcLoggingFilter
+import kr.wooco.woocobe.common.ui.filter.ReqResLoggingFilter
 import org.springframework.boot.web.servlet.FilterRegistrationBean
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.core.Ordered
 
 @Configuration
-class LoggingConfig {
+class WebFilterConfig {
     @Bean
     fun mdcLoggingFilter(): FilterRegistrationBean<MdcLoggingFilter> =
         FilterRegistrationBean<MdcLoggingFilter>().apply {
